fix: correct quiz question progression in Zustand store
- Fix nextQuestion logic to properly increment question index and update state
- Update quiz page to get questions from store instead of direct access
- Fix import issues by consolidating quiz store imports from entities
- Ensure question progression works correctly without skipping questions
