+++ src/features/quiz/index.ts
@@ -2,10 +2,3 @@ export { PopularQuiz } from "./ui/PopularQuiz";
 export { AnswersList } from "./ui/AnswersList";
 export { GameCompletionModal } from "./ui/GameCompletionModal";
 export { QuizProgress } from "./ui/QuizProgress";
-export {
-  useQuizStore,
-  useQuizQuestion,
-  useQuizAnswer,
-  useQuizProgress,
-  useQuizActions,
-} from "./model/quizStore";
